+++ datanode/gateway/graphql/mocks/mocks.go
@@ -761,6 +761,26 @@ func (mr *MockTradingDataServiceClientV2MockRecorder) GetVegaTime(arg0, arg1 int
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetVegaTime", reflect.TypeOf((*MockTradingDataServiceClientV2)(nil).GetVegaTime), varargs...)
 }
 
+// GetVotesByProposal mocks base method.
+func (m *MockTradingDataServiceClientV2) GetVotesByProposal(arg0 context.Context, arg1 *v2.GetVotesByProposalRequest, arg2 ...grpc.CallOption) (*v2.GetVotesByProposalResponse, error) {
+	m.ctrl.T.Helper()
+	varargs := []interface{}{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "GetVotesByProposal", varargs...)
+	ret0, _ := ret[0].(*v2.GetVotesByProposalResponse)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// GetVotesByProposal indicates an expected call of GetVotesByProposal.
+func (mr *MockTradingDataServiceClientV2MockRecorder) GetVotesByProposal(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetVotesByProposal", reflect.TypeOf((*MockTradingDataServiceClientV2)(nil).GetVotesByProposal), varargs...)
+}
+
 // GetWithdrawal mocks base method.
 func (m *MockTradingDataServiceClientV2) GetWithdrawal(arg0 context.Context, arg1 *v2.GetWithdrawalRequest, arg2 ...grpc.CallOption) (*v2.GetWithdrawalResponse, error) {
 	m.ctrl.T.Helper()