+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - [6527](https://github.com/vegaprotocol/vega/issues/6527) - Add `last-block` sub-command to `datanode CLI`
 
 ### 🐛 Fixes
+- [6356](https://github.com/vegaprotocol/vega/issues/6356) - When querying for proposals from `GQL` return votes.
 - [6480](https://github.com/vegaprotocol/vega/issues/6480) - Wallet `openrpc.json` is now a valid OpenRPC file
 - [6473](https://github.com/vegaprotocol/vega/issues/6473) - Infrastructure Fee Account returns error when asset is pending listing
 - [5690](https://github.com/vegaprotocol/vega/issues/5690) - Markets query now excludes rejected markets